api calls for game moved to a separate file, game server api paths moved to the router
