+++ src/services/api.ts
@@ -33,7 +33,6 @@ export const logOut = async (): Promise<IApiClientResponse> => {
   });
 };
 
-
 export const changeUserProfile = async (
   data: IApiClientResponse,
 ): Promise<IApiClientResponse> => {
@@ -61,53 +60,3 @@ export const changePasswordRequest = async (
     data,
   });
 };
-
-/*
-    GAME API
- */
-
-//TODO adapt the Api class
-const baseUrl = 'http://localhost:8081'
-
-export const relayLocalDescriptions = (peerId: string, event: string, userId: string, data: RTCSessionDescriptionInit | RTCIceCandidate) => {
-  return fetch(baseUrl + `/relay/${peerId}/${event}?user_id=${userId}`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(data)
-  });
-}
-
-export const createUser = async (username: string): Promise<string> => {
-  const res = await fetch(baseUrl + '/create', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({
-      username: username
-    })
-  });
-  const userId = await res.json();
-  return userId;
-}
-
-
-export const joinRoom = (roomId: string, userId: string) => {
-  return fetch(baseUrl + `/${roomId}/join?user_id=${userId}`, {
-    method: 'POST',
-  });
-}
-
-export const postChatMessage = (message: {content: string, username: string}) => {
-  console.log(message)
-  return fetch(baseUrl + '/message', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify(message)
-  });
-}
-
